refactor(tree): move metered execution functions to tree module
Moves the metered execution functions from ExecutorMetrics in the evm crate to EngineApiMetrics in the tree module. This consolidates execution logic directly in the tree handler for better code organization.

Changes:
- Moved metered(), execute_metered(), and metered_one() functions
- Moved MeteredStateHook struct to tree module
- Updated all call sites to use new locations
- Removed metrics feature from reth-stages dependency
- Added reth-execution-types dependency to engine/tree
- Fixed formatting issues

Closes #16247
