+++ crates/engine/tree/Cargo.toml
@@ -18,6 +18,7 @@ reth-consensus.workspace = true
 reth-db.workspace = true
 reth-engine-primitives.workspace = true
 reth-errors.workspace = true
+reth-execution-types.workspace = true
 reth-evm = { workspace = true, features = ["metrics"] }
 reth-network-p2p.workspace = true
 reth-payload-builder.workspace = true
