+++ crates/stages/stages/Cargo.toml
@@ -20,7 +20,7 @@ reth-consensus.workspace = true
 reth-db.workspace = true
 reth-db-api.workspace = true
 reth-etl.workspace = true
-reth-evm = { workspace = true, features = ["metrics"] }
+reth-evm.workspace = true
 reth-era-downloader.workspace = true
 reth-era-utils.workspace = true
 reth-era.workspace = true
